services/scheduler/pyproject.toml
[project]
name = "flights-alert-scheduler"
version = "0.1.0"
description = "Scheduler service for flights-alert"
requires-python = ">=3.13"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "openai>=1.3.0"
]
